+++ layers/__init__.py
@@ -1,2 +1,2 @@
 # Layers implemented for language modeling.
-from layers import bigram_lm, gpt, layer_utils
+from layers import bigram_lm, layer_utils, mlp, masked_multihead_attention, transformer_decoder_block, gpt
